+++ src/scenes/Languages/List/LanguageList.tsx
@@ -3,9 +3,9 @@ import { Skeleton } from '@material-ui/lab';
 import { times } from 'lodash';
 import React, { FC } from 'react';
 import { Language } from '../../../api';
-import { ContentContainer } from '../../../components/ContentContainer';
 import { useNumberFormatter } from '../../../components/Formatters';
 import { LanguageListItemCard } from '../../../components/LanguageListItemCard';
+import { ContentContainer } from '../../../components/Layout';
 import { SortButtonDialog, useSort } from '../../../components/Sort';
 import { useLanguagesQuery } from './languages.generated';
 import { LanguageSortOptions } from './LanguageSortOptions';
