+++ src/scenes/Organizations/List/OrganizationList.tsx
@@ -2,8 +2,8 @@ import { Grid, makeStyles, Typography } from '@material-ui/core';
 import { Skeleton } from '@material-ui/lab';
 import React, { FC } from 'react';
 import { Organization } from '../../../api';
-import { ContentContainer } from '../../../components/ContentContainer';
 import { useNumberFormatter } from '../../../components/Formatters';
+import { ContentContainer } from '../../../components/Layout';
 import { OrganizationListItemCard } from '../../../components/OrganizationListItemCard';
 import { SortButtonDialog, useSort } from '../../../components/Sort';
 import { listOrPlaceholders } from '../../../util';
