+++ src/scenes/Projects/List/ProjectList.tsx
@@ -2,9 +2,9 @@ import { Grid, makeStyles, Typography } from '@material-ui/core';
 import { Skeleton } from '@material-ui/lab';
 import React, { FC } from 'react';
 import { Project } from '../../../api';
-import { ContentContainer } from '../../../components/ContentContainer';
 import { FilterButtonDialog } from '../../../components/Filter';
 import { useNumberFormatter } from '../../../components/Formatters';
+import { ContentContainer } from '../../../components/Layout';
 import { ProjectListItemCard } from '../../../components/ProjectListItemCard';
 import { SortButtonDialog, useSort } from '../../../components/Sort';
 import { listOrPlaceholders } from '../../../util';
